requirements.txt
certifi==2025.1.31
Django==5.2
djangorestframework==3.16.0
fastapi==0.115.12
inference_sdk==0.48.0
langchain==0.3.25
langchain_core==0.3.65
langchain_google_genai==2.1.5
Pillow==11.2.1
psycopg2==2.9.10
psycopg2_binary==2.9.10
pydantic==2.11.5
Requests==2.32.4
stripe==12.2.0
uvicorn
dotenv
django_extensions
django-allauth
django-cors-headers
jwt
requests-oauthlib